+++ src/main/java/com/ifood/quarkus/crud/entity/Localizacao.java
@@ -1,5 +1,7 @@
 package com.ifood.quarkus.crud.entity;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -8,14 +10,15 @@ import javax.persistence.Id;
 import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
 
 @Entity
+ 
 public class Localizacao extends PanacheEntityBase {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	public Long id;
 
-	public Double latitude;
+	public BigDecimal latitude;
 
-	public Double longitude;
+	public BigDecimal longitude;
 
 }
